Mark a room as joined by nulling ends_at in database
As discovered during testing, it is possible for a client to open a game
via an osu-web api call but never join it (due to client crash, for
instance).

To address this, osu-web will now give a host 30 seconds to join the
match before it is automatically closed. This is done via ends_at (see
https://github.com/ppy/osu-web/pull/7013). We now need to resolve rooms
as "joined" by setting this field to null.