Fix intermingling multiple token types
The replay upload flow and the user statistics update flow would rely on
the score token sent by client to spectator server in
`BeginPlaySession()`, in order to unmap the score token to a solo score
ID when processing finishes.

However, spectator server thus far has not been aware that there are
actually _two_ types of token - and to be fair, client gave it no
practical way to actually know this. In solo, the token ID comes from
`solo_score_tokens`, and in multiplayer, the token ID comes from
`multiplayer_score_links`. Spectator server has however been treating
both types as the solo score type. In particular, this could mean that
the replay recording flow would royally confuse things and unmap the
token to a score ID wrong, and as such, upload an entirely misattributed
replay.

To resolve, leverage the updated client contract which now passes a
`ScoreToken` with an explicit type rather than a `long` to do the right
thing for both types of token.