Adjust end match database query in line with multiplayer score storage changes
With https://github.com/ppy/osu-web/pull/10437, scores go into the
`multiplayer_score_links` table, rather than the `multiplayer_scores`
table.

The primary case impacted by this change is when a user suddenly
disconnects from an open room in which they are alone. In such a case,
without this change, the `expired` check would fail (since the item, at
the time of the sole player's exit, is not yet finished / expired), and
as such due to not checking the new multiplayer score table, spectator
server would consider the relevant playlist item unplayed and delete it,
leaving behind a `multiplayer_score_links` record with
`playlist_item_id` leading into the void.