Upgrade to .net 8
The majority of the overhead we're seeing is system%. Hardly any of what
we are doing is causing overhead. Given the amount of optimisation done
between net6 and net8, this should give us some headroom.

Willing to take a gamble on this one under the premise that we can
rollback if required. Will test on staging for efficiency.

https://github.com/dotnet/AspNetCore.Docs/issues/17880