Skip score uploads when nothing was hit in score
The client will not submit a play if hit statistics indicate that they
never hit anything:

	https://github.com/ppy/osu/blob/a47ccb8edd2392258b6b7e176b222a9ecd511fc0/osu.Game/Screens/Play/SubmittingPlayer.cs#L281

This check was not exercised spectator server-side, which meant that
scores with nothing hit would get completed and enqueued for upload that
would never succeed (because the score would never be submitted anyway).
To save on some processing, just port the same check server-side to skip
upload if nothing was hit.

I'm hoping the effects of this is going to be the decrease of the score
timeouts monitoring on sentry & datadog closer to real levels (see
discussion in https://github.com/ppy/osu-server-spectator/issues/220).