Remove spectator tracking
In discussion it appears that the consensus is that the player should
not be able to spectate offline players, and as such the server is
allowed to purge the list of spectators on disconnection of the
streamer. This in turn obviates the need of doing any external tracking
of who's spectating whom in the hub.