Fix aborts not being logged to events if they occur before play starts
Fell out of investigating more cases of multiplayer events rows with
missing playlists items:

https://discord.com/channels/90072389919997952/1327149041511043134/1428261428045021255

The root failure here was actually fixed in
f42fd8f3799a2cc6ab4b3dcb01393447d3eb3909 via the addition of the
following call:

https://github.com/ppy/osu-server-spectator/blob/4f585909ead4bc89fdfdedb237a8c3fff6be8c26/osu.Server.Spectator/Hubs/Multiplayer/MultiplayerHubContext.cs#L337

without which, the following sequence of events was possible:

- Gameplay start is requested, which results in the `game_started` event
  being logged
- Gameplay is aborted during load
- Due to the lack of the aforementioned call, the playlist item is *not*
  expired
- And thus the item can be then arbitrarily edited or even removed,
  causing the failure.

This can be semi-confirmed by cherry-picking
983a0d0fe9f4baeb5a960cb76e7adadd2f02623f containing the relevant test
coverage onto 2025.801.0 which is currently live.

The remaining issue is a lack of logging of the abort in the same
scenario.