Fix replay upload hard failures not running relevant cleanup logic
As pointed out in
https://discord.com/channels/90072389919997952/983550677794050108/1429504084813152336.

Also includes a forgotten `CancellationTokenSource` disposal, which is
semi-relevant because that would lead to not disposing some `Timer`s
internal to the token source. Probably still not going to lead to actual
leaks because I expect the timer to be backed by `WaitHandle`s, but
better safe than sorry, probably.