Do not allow removal of playlist items with issued score tokens
@nanaya has pointed out in
https://discord.com/channels/90072389919997952/1327149041511043134/1448230657867120671
that we apparently have cases in the database wherein
`multiplayer_score_links` rows are associated with playlist items that
no longer exist.

The playlist item ID in `multiplayer_score_links` comes directly from
the score token:

	https://github.com/ppy/osu-web/blob/192447ea92b4dbf8990533710c518df876c770ad/app/Models/Multiplayer/ScoreLink.php#L34

So the logical temporal conclusion is that the playlist item must have
been removed after there have been score tokens issued for it.

The only operation left I'm aware of that is able to straight up delete
playlist items from database is the `RemovePlaylistItem()` operation
triggered by clients. So this commit adds a guard whose intention is to
make that operation fail if there are already score tokens for said
playlist item in the database.

You will inevitably ask, "but spaceman, how can this happen in the first
place?" My answer to that is, uhh, *blows raspberry*. I don't know. I
don't see obvious deficiencies in locking or anything of the sort. May
be a network race for all I know.

Then you will ask, "why PR this then?" Well I consider playlist items
*with scores attached* getting dumpstered as pants-on-fire breakage,
that even if rare, needs to be plugged ASAP, no ifs, ands, or buts.

Any follow-ups can happen if this does fire in production and maybe we
get some extra data points as to how and when this can happen off of
that.